fix: minor restructure of server moved ApplicationReview into Evalution Module moved core.domain.constants from to core.constants
