+++ src/main/java/de/tum/cit/aet/evaluation/domain/ApplicationReview.java
@@ -1,5 +1,6 @@
-package de.tum.cit.aet.application.domain;
+package de.tum.cit.aet.evaluation.domain;
 
+import de.tum.cit.aet.application.domain.Application;
 import de.tum.cit.aet.usermanagement.domain.User;
 import jakarta.persistence.*;
 import java.time.Instant;
